backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
numpy>=1.24.0
qdrant-client>=1.7.0
transformers>=4.35.0
torch>=2.0.0
tree-sitter>=0.20.0
tree-sitter-languages>=1.8.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
requests>=2.31.0
python-dotenv>=1.0.0
slowapi>=0.1.9
PyJWT>=2.8.0
python-jose[cryptography]>=3.3.0
